@@ -4,6 +4,6 @@ description: A library for generating fake data. faker is heavily inspired by th
 homepage: https://github.com/drager/faker
 author: Jesper Håkansson <jesper@jesperh.se>
 environment:
-  sdk: '>=2.0.0 <3.0.0'
+  sdk: '>=2.12.0-0 <3.0.0'
 dev_dependencies:
   test: ^1.2.0
